split cancel and delete to difrent call by parameters

+++ WebAPI/Clients/ConditionalAccessClient.cs
@@ -1685,7 +1685,7 @@ namespace WebAPI.Clients
             return seriesRecording;
         }
 
-        internal KalturaSeriesRecording DeleteSeriesRecord(int groupId, string userId, long domainId, long id, long epgId = 0, long seasonNumber = 0)
+        internal KalturaSeriesRecording DeleteSeriesRecord(int groupId, string userId, long domainId, long id, long epgId = 0, int seasonNumber = 0)
         {
             KalturaSeriesRecording seriesRecording = null;
             SeriesRecording response = null;
